# Conflicts:
#	app/src/main/java/com/fieldbook/tracker/offbeat/traits/formats/contracts/UsbCameraFormat.kt

+++ app/src/main/java/com/fieldbook/tracker/traits/AbstractCameraTrait.kt
@@ -7,6 +7,7 @@ import android.content.Intent
 import android.content.SharedPreferences
 import android.graphics.Bitmap
 import android.graphics.BitmapFactory
+import android.graphics.Matrix
 import android.graphics.Point
 import android.net.Uri
 import android.os.Build
@@ -15,11 +16,13 @@ import android.util.AttributeSet
 import android.util.Log
 import android.widget.ImageView
 import android.widget.Toast
+import androidx.camera.view.PreviewView
 import androidx.documentfile.provider.DocumentFile
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.fieldbook.tracker.R
 import com.fieldbook.tracker.activities.CollectActivity
+import com.fieldbook.tracker.adapters.ImageAdapter
 import com.fieldbook.tracker.database.models.ObservationModel
 import com.fieldbook.tracker.objects.RangeObject
 import com.fieldbook.tracker.preferences.GeneralKeys
@@ -31,7 +34,6 @@ import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
-import org.phenoapps.adapters.ImageAdapter
 import org.phenoapps.androidlibrary.Utils
 import java.io.FileNotFoundException
 import javax.inject.Inject
@@ -46,16 +48,26 @@ abstract class AbstractCameraTrait :
 
     companion object {
         const val TAG = "Camera"
+        const val TEMPORARY_IMAGE_NAME = "temp.jpg"
+    }
+
+    enum class SaveState {
+        SINGLE_SHOT, NEW, SAVING, COMPLETE
     }
 
     protected var activity: Activity? = null
     protected var connectBtn: FloatingActionButton? = null
     protected var captureBtn: FloatingActionButton? = null
+    protected var settingsBtn: FloatingActionButton? = null
+    protected var previewView: PreviewView? = null
     protected var imageView: ImageView? = null
     protected var recyclerView: RecyclerView? = null
 
     protected val background = CoroutineScope(Dispatchers.IO)
     protected val ui = CoroutineScope(Dispatchers.Main)
+
+    open val thumbnailWidth = 256
+    open val thumbnailHeight = 256
 
     constructor(context: Context?) : super(context)
     constructor(context: Context?, attrs: AttributeSet?) : super(context, attrs)
@@ -89,18 +101,78 @@ abstract class AbstractCameraTrait :
         captureBtn = act.findViewById(R.id.camera_fragment_capture_btn)
         imageView = act.findViewById(R.id.trait_camera_iv)
         recyclerView = act.findViewById(R.id.camera_fragment_rv)
+        settingsBtn = act.findViewById(R.id.camera_fragment_settings_btn)
+        previewView = act.findViewById(R.id.trait_camera_pv)
 
         recyclerView?.adapter = ImageAdapter(this)
 
         activity = act
 
+        settingsBtn?.setOnClickListener {
+            showSettings()
+        }
+    }
+
+    abstract fun showSettings()
+
+    abstract fun onSettingsChanged()
+
+    open fun getThumbnailSize() = Point(thumbnailWidth, thumbnailHeight)
+
+    protected fun saveJpegToStorage(
+        format: String,
+        data: ByteArray,
+        obsUnit: RangeObject,
+        saveTime: String,
+        saveState: SaveState
+    ) {
+
+        saveToStorage(format, obsUnit, saveTime, saveState) { uri ->
+
+            context.contentResolver.openOutputStream(uri, "wa")?.use { output ->
+
+                output.write(data)
+
+            }
+        }
     }
 
     protected fun saveBitmapToStorage(format: String, bmp: Bitmap, obsUnit: RangeObject) {
 
+        saveToStorage(format, obsUnit, saveTime = Utils.getDateTime(), saveState = SaveState.NEW) { uri ->
+
+            context.contentResolver.openOutputStream(uri)?.let { output ->
+
+                bmp.compress(Bitmap.CompressFormat.JPEG, 80, output)
+
+            }
+        }
+    }
+
+    private fun writeExif(file: DocumentFile) {
+
+        //if sdk > 24, can write exif information to the image
+        //goal is to encode observation variable model into the user comments
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
+
+            ExifUtil.saveJsonToExif(
+                context,
+                currentTrait,
+                file.uri
+            )
+        }
+    }
+
+    private fun saveToStorage(
+        format: String,
+        obsUnit: RangeObject,
+        saveTime: String,
+        saveState: SaveState,
+        saver: (Uri) -> Unit
+    ) {
+
         val plot = obsUnit.plot_id
         val studyId = collectActivity.studyId
-        val time = Utils.getDateTime()
         val person = (activity as? CollectActivity)?.person
         val location = (activity as? CollectActivity)?.locationByPreferences
 
@@ -115,15 +187,15 @@ abstract class AbstractCameraTrait :
 
                 //get the bitmap from the texture view, only use it if its not null
 
+                val name = "${sanitizedTraitName}_${plot}_$saveTime.jpg"
+
                 DocumentTreeUtil.getFieldMediaDirectory(context, sanitizedTraitName)?.let { dir ->
 
-                    val name = "${sanitizedTraitName}_${plot}_$time.jpg"
+                    if (saveState == SaveState.NEW || saveState == SaveState.SINGLE_SHOT) {
 
-                    dir.createFile("*/*", name)?.let { file ->
+                        dir.createFile("*/*", name)?.let { file ->
 
-                        context.contentResolver.openOutputStream(file.uri)?.let { output ->
-
-                            bmp.compress(Bitmap.CompressFormat.JPEG, 80, output)
+                            saver.invoke(file.uri)
 
                             database.insertObservation(
                                 plot, traitDbId, format, file.uri.toString(),
@@ -134,23 +206,41 @@ abstract class AbstractCameraTrait :
                                 null
                             )
 
-                            //if sdk > 24, can write exif information to the image
-                            //goal is to encode observation variable model into the user comments
-                            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
+                            if (saveState == SaveState.SINGLE_SHOT) {
 
-                                ExifUtil.saveJsonToExif(
-                                    context,
-                                    currentTrait,
-                                    file.uri
-                                )
+                                writeExif(file)
+
+                                notifyItemInserted()
+                            }
+                        }
+
+                    } else if (saveState in setOf(SaveState.SAVING, SaveState.COMPLETE)) {
+
+                        dir.findFile(name)?.let { file ->
+
+                            if (saveState == SaveState.COMPLETE) {
+
+                                writeExif(file)
+
+                                notifyItemInserted()
+
+                            } else {
+
+                                saver.invoke(file.uri)
+
                             }
                         }
                     }
-
-                    loadAdapterItems()
-
                 }
             }
+        }
+    }
+
+    private fun notifyItemInserted() {
+
+        ui.launch {
+
+            loadAdapterItems()
         }
     }
 
@@ -211,7 +301,6 @@ abstract class AbstractCameraTrait :
 
             }, 500L)
 
-
         } catch (e: Exception) {
 
             e.printStackTrace()
@@ -221,9 +310,16 @@ abstract class AbstractCameraTrait :
 
     private fun loadAdapterItems() {
 
+        val studyId = collectActivity.studyId
+        val plot = collectActivity.observationUnit
+        val traitId = currentTrait.id
+        val rep = collectActivity.rep
+
         background.launch {
 
             val thumbnailModels = getImageObservations().mapNotNull {
+
+                val brapiSynced = database.isBrapiSynced(studyId, plot, traitId, rep)
 
                 var model: ImageAdapter.Model? = null
 
@@ -231,10 +327,26 @@ abstract class AbstractCameraTrait :
 
                     DocumentsContract.getDocumentThumbnail(
                         context.contentResolver,
-                        Uri.parse(it.value), Point(256, 256), null
+                        Uri.parse(it.value), getThumbnailSize(), null
                     )?.let { bmp ->
 
-                        model = ImageAdapter.Model(it.value, bmp)
+                        model = ImageAdapter.Model(it.value, if (bmp.height > bmp.width) {
+
+                            val matrix = Matrix()
+
+                            matrix.postRotate(90f)
+
+                            Bitmap.createBitmap(
+                                bmp,
+                                0,
+                                0,
+                                bmp.width,
+                                bmp.height,
+                                matrix,
+                                true
+                            )
+
+                        } else bmp, brapiSynced)
 
                     }
 
@@ -266,13 +378,76 @@ abstract class AbstractCameraTrait :
         return database.getAllObservations(studyId, plot, traitDbId)
     }
 
-    private fun deleteItem(model: ImageAdapter.Model) {
+    /**
+     * Uses details parameter to check if the number of photos has been reached.
+     */
+    protected fun checkPictureLimit(): Boolean {
+
+        try {
+
+            val m: Int = try {
+
+                currentTrait.details.toInt()
+
+            } catch (n: Exception) {
+
+                0
+            }
+
+            // Do not take photos if limit is reached
+            return m == 0 || getImageObservations().size < m
+
+        } catch (e: Exception) {
+
+            e.printStackTrace()
+
+            com.fieldbook.tracker.utilities.Utils.makeToast(context, context.getString(R.string.trait_error_hardware_missing))
+        }
+
+        return false
+    }
+
+    private fun updateObservationValueToNa(observation: ObservationModel) {
+
+        database.updateObservation(observation.also {
+            it.value = "NA"
+        })
+    }
+
+    private fun deleteImage(image: DocumentFile) {
 
         val studyId = prefs.getInt(GeneralKeys.SELECTED_FIELD_ID, 0).toString()
 
         val plot = currentRange.plot_id
 
         val traitDbId = currentTrait.id
+
+        val result = image.delete()
+
+        if (result) {
+
+            database.deleteTraitByValue(
+                studyId,
+                plot,
+                traitDbId,
+                image.uri.toString()
+            )
+
+        } else {
+
+            collectActivity.runOnUiThread {
+
+                Toast.makeText(
+                    context,
+                    R.string.photo_failed_to_delete,
+                    Toast.LENGTH_SHORT
+                ).show()
+
+            }
+        }
+    }
+
+    private fun deleteItem(model: ImageAdapter.Model) {
 
         getImageObservations().firstOrNull { it.value == model.uri }?.let { observation ->
 
@@ -281,35 +456,16 @@ abstract class AbstractCameraTrait :
                 DocumentFile.fromSingleUri(context, Uri.parse(observation.value))
                     ?.let { image ->
 
-                        val result = image.delete()
+                        if (model.brapiSynced) {
 
-                        if (result) {
+                            image.delete()
 
-                            database.deleteTraitByValue(
-                                studyId,
-                                plot,
-                                traitDbId,
-                                image.uri.toString()
-                            )
+                            updateObservationValueToNa(observation)
 
-                            //val index = (recyclerView?.adapter as ImageAdapter).currentList.indexOf(model)
+                        } else deleteImage(image)
 
-                            //recyclerView?.adapter as ImageAdapter).notifyItemRemoved(index)
+                        loadAdapterItems()
 
-                            loadAdapterItems()
-
-                        } else {
-
-                            collectActivity.runOnUiThread {
-
-                                Toast.makeText(
-                                    context,
-                                    R.string.photo_failed_to_delete,
-                                    Toast.LENGTH_SHORT
-                                ).show()
-
-                            }
-                        }
                     }
 
             } catch (e: Exception) {
@@ -317,6 +473,9 @@ abstract class AbstractCameraTrait :
                 Log.e(TAG, "Failed to delete images.", e)
 
             }
+
+            (context as CollectActivity).refreshRepeatedValuesToolbarIndicator()
+
         }
     }
 
@@ -329,6 +488,7 @@ abstract class AbstractCameraTrait :
                 DocumentFile.fromSingleUri(context, Uri.parse(observation.value))?.let { image ->
 
                     activity?.startActivity(Intent(Intent.ACTION_VIEW, image.uri).also {
+                        it.setDataAndType(image.uri, "image/*")
                         it.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION)
                     })
                 }
@@ -346,4 +506,3 @@ abstract class AbstractCameraTrait :
         (context as CollectActivity).traitLockData()
     }
 }
-
